# job-info: index watchers for quick cancels