--- a/src/modules/job-info/guest_watch.c
+++ b/src/modules/job-info/guest_watch.c
@@ -120,6 +120,8 @@ struct guest_watch_ctx {
     int guest_offset;
     /* data from main namespace */
     int main_offset;
+
+    void *handle;               /* zlistx_t handle */
 };
 
 static int get_main_eventlog (struct guest_watch_ctx *gw);
@@ -144,6 +146,15 @@ static void guest_watch_ctx_destroy (void *data)
         flux_future_destroy (gw->main_namespace_watch_f);
         free (gw);
         errno = save_errno;
+    }
+}
+
+/* zlistx_destructor_fn */
+static void guest_watch_ctx_destroy_wrapper (void **data)
+{
+    if (data) {
+        guest_watch_ctx_destroy (*data);
+        *data = NULL;
     }
 }
 
@@ -360,8 +371,8 @@ error:
         flux_log_error (ctx->h, "%s: flux_respond_error", __FUNCTION__);
 done:
     /* flux future destroyed in guest_watch_ctx_destroy, which is
-     * called via zlist_remove() */
-    zlist_remove (ctx->guest_watchers, gw);
+     * called via zlistx_delete() */
+    zlistx_delete (ctx->guest_watchers, gw->handle);
 }
 
 static int wait_guest_namespace (struct guest_watch_ctx *gw)
@@ -523,8 +534,8 @@ error:
         flux_log_error (ctx->h, "%s: flux_respond_error", __FUNCTION__);
 cleanup:
     /* flux future destroyed in guest_watch_ctx_destroy, which is
-     * called via zlist_remove() */
-    zlist_remove (ctx->guest_watchers, gw);
+     * called via zlistx_delete() */
+    zlistx_delete (ctx->guest_watchers, gw->handle);
 }
 
 static int guest_namespace_watch (struct guest_watch_ctx *gw)
@@ -646,8 +657,8 @@ error:
         flux_log_error (ctx->h, "%s: flux_respond_error", __FUNCTION__);
 cleanup:
     /* flux future destroyed in guest_watch_ctx_destroy, which is called
-     * via zlist_remove() */
-    zlist_remove (ctx->guest_watchers, gw);
+     * via zlistx_delete() */
+    zlistx_delete (ctx->guest_watchers, gw->handle);
 }
 
 /* must prefix "guest." back to path when watching in main KVS
@@ -767,8 +778,8 @@ error:
         flux_log_error (ctx->h, "%s: flux_respond_error", __FUNCTION__);
 cleanup:
     /* flux future destroyed in guest_watch_ctx_destroy, which is called
-     * via zlist_remove() */
-    zlist_remove (ctx->guest_watchers, gw);
+     * via zlistx_delete() */
+    zlistx_delete (ctx->guest_watchers, gw->handle);
 }
 
 int guest_watch (struct info_ctx *ctx,
@@ -785,11 +796,11 @@ int guest_watch (struct info_ctx *ctx,
     if (get_main_eventlog (gw) < 0)
         goto error;
 
-    if (zlist_append (ctx->guest_watchers, gw) < 0) {
-        flux_log_error (ctx->h, "%s: zlist_append", __FUNCTION__);
+    if (!(gw->handle = zlistx_add_end (ctx->guest_watchers, gw))) {
+        errno = ENOMEM;
+        flux_log_error (ctx->h, "%s: zlistx_add_end", __FUNCTION__);
         goto error;
     }
-    zlist_freefn (ctx->guest_watchers, gw, guest_watch_ctx_destroy, true);
     gw = NULL;
 
     return 0;
@@ -821,17 +832,18 @@ void guest_watchers_cancel (struct info_ctx *ctx,
 {
     struct guest_watch_ctx *gw;
 
-    gw = zlist_first (ctx->guest_watchers);
+    gw = zlistx_first (ctx->guest_watchers);
     while (gw) {
         guest_watch_cancel (ctx, gw, msg, cancel);
-        gw = zlist_next (ctx->guest_watchers);
+        gw = zlistx_next (ctx->guest_watchers);
     }
 }
 
 int guest_watch_setup (struct info_ctx *ctx)
 {
-    if (!(ctx->guest_watchers = zlist_new ()))
+    if (!(ctx->guest_watchers = zlistx_new ()))
         return -1;
+    zlistx_set_destructor (ctx->guest_watchers, guest_watch_ctx_destroy_wrapper);
     return 0;
 }
 
@@ -840,7 +852,8 @@ void guest_watch_cleanup (struct info_ctx *ctx)
     if (ctx->guest_watchers) {
         struct guest_watch_ctx *gw;
 
-        while ((gw = zlist_pop (ctx->guest_watchers))) {
+        while (zlistx_first (ctx->guest_watchers)) {
+            gw = zlistx_detach_cur (ctx->guest_watchers);
             send_eventlog_watch_cancel (gw, NULL, false);
 
             if (flux_respond_error (ctx->h, gw->msg, ENOSYS, NULL) < 0)
@@ -848,7 +861,7 @@ void guest_watch_cleanup (struct info_ctx *ctx)
                                 __FUNCTION__);
             guest_watch_ctx_destroy (gw);
         }
-        zlist_destroy (&ctx->guest_watchers);
+        zlistx_destroy (&ctx->guest_watchers);
         ctx->guest_watchers = NULL;
     }
 }