--- a/src/modules/job-info/job-info.h
+++ b/src/modules/job-info/job-info.h
@@ -19,13 +19,21 @@
 
 #define OWNER_LRU_MAXSIZE 1000
 
+/* N.B. zlistx_t is predominantly use for storage b/c we need to
+ * iterate and remove entries while iterating.  This cannot be done
+ * with a zhashx_t unless we use the relatively costly zhashx_keys().
+ * So a number of lists are supplemented with a hash for when a
+ * quick lookup would be advantageous.
+ */
 struct info_ctx {
     flux_t *h;
     flux_msg_handler_t **handlers;
     lru_cache_t *owner_lru; /* jobid -> owner LRU */
     zlistx_t *lookups;
     zlistx_t *watchers;
+    zhashx_t *watchers_matchtags; /* matchtag + uuid -> watcher */
     zlistx_t *guest_watchers;
+    zhashx_t *guest_watchers_matchtags; /* matchtag + uuid -> guest_watcher */
     zlistx_t *update_watchers;
     zhashx_t *index_uw;        /* jobid + key -> update_watcher lookup */
 };