--- a/src/modules/job-info/update.h
+++ b/src/modules/job-info/update.h
@@ -38,6 +38,8 @@ void update_watchers_cancel (struct info_ctx *ctx,
                              const flux_msg_t *msg,
                              bool cancel);
 
+int update_watch_setup (struct info_ctx *ctx);
+
 void update_watch_cleanup (struct info_ctx *ctx);
 
 int update_watch_count (struct info_ctx *ctx);