Problem: In watchers_cancel() and guest_watchers_cancel(), every
watcher is iterated over to try and match a cancellation request.
This can slow down performance when the list grows larger and larger
over time.  For example, when monitoring jobs (i.e. watching
eventlogs of jobs) this can lead to very long lists.

Solution: Create an index that will map request uuids & matchtags
to their respective watchers, which will allow quick lookups for
cancellation requests.

---

With more and more performance fixes up in master, the gap between "master" and my PRs shrinks, we gotta get to 64K jobs to show the big improvement.

Current master

```
Running throughput.py 2048 jobs per iteration, 1 iterations
throughput:     198.6 job/s (script: 192.1 job/s)
1: 11 seconds
Total of 11 seconds for everything
Running throughput.py 4096 jobs per iteration, 1 iterations
throughput:     183.4 job/s (script: 181.2 job/s)
1: 23 seconds
Total of 23 seconds for everything
Running throughput.py 8192 jobs per iteration, 1 iterations
throughput:     180.2 job/s (script: 178.7 job/s)
1: 46 seconds
Total of 46 seconds for everything
Running throughput.py 16384 jobs per iteration, 1 iterations
throughput:     162.8 job/s (script: 162.1 job/s)
1: 101 seconds
Total of 101 seconds for everything
Running throughput.py 32768 jobs per iteration, 1 iterations
throughput:     131.9 job/s (script: 131.7 job/s)
1: 249 seconds
Total of 249 seconds for everything
Running throughput.py 65536 jobs per iteration, 1 iterations
throughput:     43.6 job/s (script:  43.6 job/s)
1: 1503 seconds
Total of 1503 seconds for everything

```

this branch

```
Running throughput.py 2048 jobs per iteration, 1 iterations
throughput:     197.4 job/s (script: 191.2 job/s)
1: 11 seconds
Total of 11 seconds for everything
Running throughput.py 4096 jobs per iteration, 1 iterations
throughput:     180.3 job/s (script: 178.2 job/s)
1: 23 seconds
Total of 23 seconds for everything
Running throughput.py 8192 jobs per iteration, 1 iterations
throughput:     176.0 job/s (script: 174.5 job/s)
1: 48 seconds
Total of 48 seconds for everything
Running throughput.py 16384 jobs per iteration, 1 iterations
throughput:     161.6 job/s (script: 161.0 job/s)
1: 102 seconds
Total of 102 seconds for everything
Running throughput.py 32768 jobs per iteration, 1 iterations
throughput:     137.0 job/s (script: 136.7 job/s)
1: 240 seconds
Total of 240 seconds for everything
Running throughput.py 65536 jobs per iteration, 1 iterations
throughput:     68.4 job/s (script:  68.4 job/s)
1: 960 seconds
Total of 960 seconds for everything
```

so through 32K jobs, not too much of a difference, but a significant difference at 64K.